Added logic in UserDetailedScreen, Firestore user data creation, authentication state check, and DeciderScreen as start destination (#10)
Added the functionality to create the user data in firestore upon first google login via UserDetailedScreen. Added the check of new user and existing user and behaves according to it. Removed the dynamic startDestination variable in navHostController and Added a DeciderScreen as a startDestination which will be used as SplashScreen in future, decides which screen to open based on AuthState.